requirements.txt
agno==2.0.6
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.8.1
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.1
distro==1.9.0
Django==5.1.4
django-bootstrap5==24.3
django-resized==1.0.3
djangorestframework==3.15.2
dnspython==2.8.0
docstring_parser==0.17.0
email-validator==2.3.0
gitdb==4.0.12
GitPython==3.1.45
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
kombu==5.5.4
markdown-it-py==4.0.0
mdurl==0.1.2
openai==1.108.0
packaging==25.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.1.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.7
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.0.1
rich==14.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlparse==0.5.3
tenacity==9.1.2
tqdm==4.67.1
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
